feat: move control context and add color scale
