+++ src/frontend/acc/src/components/results/controls/controls-wrapper.tsx
@@ -1,12 +1,9 @@
 import { Busy } from "@acc/components/shared/busy";
 import { Card } from "@acc/components/ui/card";
-import { ControlsContextProvider } from "@acc/lib/contexts/controls-context";
 import MolstarPartialCharges from "@acc/lib/viewer/viewer";
-import { HTMLAttributes, useState } from "react";
+import { HTMLAttributes } from "react";
 
-import { MolstarColoringType } from "./coloring-controls";
 import { Controls } from "./controls";
-import { MolstarViewType } from "./view-controls";
 
 export type ControlsWrapperProps = {
   computationId: string;
@@ -14,61 +11,21 @@ export type ControlsWrapperProps = {
   molecules: string[];
 } & HTMLAttributes<HTMLElement>;
 
-// Overriding the default view type for specific examples
-const exampleOverrides: Record<string, { viewType: MolstarViewType }> = {
-  "examples/pore": {
-    viewType: "surface",
-  },
-  "examples/oseltamivir": {
-    viewType: "cartoon",
-  },
-};
-
 export const ControlsWrapper = ({
   computationId,
   molstar,
   molecules,
 }: ControlsWrapperProps) => {
-  const [currentTypeId, setCurrentTypeId] = useState<number>(1);
-  const [structure, setStructure] = useState<string>(molecules[0]);
-  const [coloringType, setColoringType] =
-    useState<MolstarColoringType>("charges-relative");
-  const [maxValue, setMaxValue] = useState<number>(0);
-  const [viewType, setViewType] = useState<MolstarViewType>(() => {
-    if (exampleOverrides?.[computationId]?.viewType) {
-      return exampleOverrides[computationId].viewType;
-    }
-    return molstar.type.isDefaultApplicable() ? "cartoon" : "balls-and-sticks";
-  });
-  const [methodNames, setMethodNames] = useState<(string | undefined)[]>([]);
-
   return (
     <Card className="w-4/5 rounded-none mx-auto p-4 max-w-content mt-4 flex flex-col relative">
-      <ControlsContextProvider
-        value={{
-          currentTypeId,
-          setCurrentTypeId,
-          coloringType,
-          setColoringType,
-          maxValue,
-          setMaxValue,
-          structure,
-          setStructure,
-          viewType,
-          setViewType,
-          methodNames,
-          setMethodNames,
-        }}
-      >
-        <Busy isBusy={!molstar} />
-        {molstar && (
-          <Controls
-            computationId={computationId}
-            molstar={molstar}
-            molecules={molecules}
-          />
-        )}
-      </ControlsContextProvider>
+      <Busy isBusy={!molstar} />
+      {molstar && (
+        <Controls
+          computationId={computationId}
+          molstar={molstar}
+          molecules={molecules}
+        />
+      )}
     </Card>
   );
 };
