+++ src/frontend/acc/src/components/results/molstar-wrapper.tsx
@@ -1,6 +1,6 @@
 import MolstarPartialCharges from "@acc/lib/viewer/viewer";
 import { PluginUIContext } from "molstar/lib/commonjs/mol-plugin-ui/context";
-import { HTMLAttributes, useEffect, useState } from "react";
+import { HTMLAttributes, useEffect, useRef, useState } from "react";
 
 import { Busy, BusySize } from "../shared/busy";
 import { MolstarViewer } from "./molstar";
@@ -9,9 +9,13 @@ export type MolstarWrapperProps = {
   setMolstar: React.Dispatch<
     React.SetStateAction<MolstarPartialCharges | undefined>
   >;
+  maxCharge?: number;
 } & HTMLAttributes<HTMLElement>;
 
-export const MolstarWrapper = ({ setMolstar }: MolstarWrapperProps) => {
+export const MolstarWrapper = ({
+  setMolstar,
+  maxCharge,
+}: MolstarWrapperProps) => {
   const [plugin, setPlugin] = useState<PluginUIContext | undefined>();
 
   const setup = async () => {
@@ -28,7 +32,7 @@ export const MolstarWrapper = ({ setMolstar }: MolstarWrapperProps) => {
   return (
     <div className="relative w-4/5 mx-auto">
       <Busy isBusy={!plugin} size={BusySize.Big} />
-      {plugin && <MolstarViewer plugin={plugin} />}
+      {plugin && <MolstarViewer maxCharge={maxCharge} plugin={plugin} />}
     </div>
   );
 };
